requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pymupdf==1.24.10
rank_bm25==0.2.2
requests==2.31.0
tenacity==9.0.0
python-dotenv==1.0.1